requirements.txt
flask
pytest
requests
pymysql
flask-sqlalchemy
pyjwt
werkzeug
pesel
xlsxwriter
faker
pylightxl
coloredlogs
ruamel.yaml